fatmug_project/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.11
django-environ==0.11.2
djangorestframework==3.15.1
mysqlclient==2.2.4
sqlparse==0.5.0
typing-extensions==4.11.0
